+++ CHANGELOG.md
@@ -9,12 +9,14 @@ CHANGELOG
 - Add support for custom actions on login nodes.
 - Allow DCV connection on login nodes.
 - Add support for ap-southeast-3 region.
+- Add security groups to login node network load balancer.
+- Add `AllowedIps` configuration for login nodes.
 
 **BUG FIXES**
 - Fix validator `EfaPlacementGroupValidator` so that it does not suggest to configure a Placement Group when Capacity Blocks are used.
 - Fix sporadic cluster creation failures with managed FSx for Lustre.
+- Fix issue with login nodes being marked unhealthy when restricting SSH access.
 - Fix `retrieve_supported_regions` so that it can get the correct S3 url.
-
 3.10.1
 ------